+++ docs/source/handlers.rst
@@ -161,7 +161,7 @@ For example:
 PubSub
 ^^^^^^
 
-You can trigger endpoints from pubsub using the ``@app.topic(...)`` decorator. All that is required is the topic name. You can optionally 
+You can trigger endpoints from pubsub using the ``@app.pubsub_subscription(...)`` decorator. All that is required is the topic name. You can optionally 
 provide an attribute dictionary which will only trigger the function if the pubsub message attributes matches those defined in the decorator.
 If using cloudrun backend or `use_subscription=true` the attributes will be created as a filter on the subscription itself. You can also pass in 
 a custom `filter` as well. Note that filters are not able to be modified once they are applied to a subscription. 
@@ -175,34 +175,34 @@ Example usage:
 .. code:: python 
 
     # pubsub topic
-    @app.topic('test')
+    @app.pubsub_subscription('test')
     def topic(data):
         app.log.info(data)
         return 
 
     # pubsub topic with matching message attributes
-    @app.topic('test', attributes={'key': 'value'})
+    @app.pubsub_subscription('test', attributes={'key': 'value'})
     def home2(data):
         app.log.info(data)
         return 
 
     # pubsub topic in a different project
-    @app.topic('test', project="CROSS_PROJECT")
+    @app.pubsub_subscription('test', project="CROSS_PROJECT")
     def cross_project(data):
         return 
 
     # create a pubsub subscription instead of pubsub triggered function
-    @app.topic('test', use_subscription=True)
+    @app.pubsub_subscription('test', use_subscription=True)
     def pubsub_subscription_use_subscription(data):
         return 
 
     # create a pubsub subscription instead of pubsub triggered function and add filter
-    @app.topic('test', use_subscription=True, filter='attributes.name = "com" AND -attributes:"iana.org/language_tag"')
+    @app.pubsub_subscription('test', use_subscription=True, filter='attributes.name = "com" AND -attributes:"iana.org/language_tag"')
     def pubsub_subscription_filter(data):
         return 
 
     # switching the pubsub topic to a different project requires force_update, since it requires the subscription to be recreated
-    @app.topic('test', project="NEW_CROSS_PROJECT", force_update=True)
+    @app.pubsub_subscription('test', project="NEW_CROSS_PROJECT", force_update=True)
     def cross_project(data):
         return
