+++ goblet/decorators.py
@@ -13,6 +13,7 @@ from goblet.backends.cloudrun import CloudRun
 from goblet.infrastructures.redis import Redis
 from goblet.infrastructures.vpcconnector import VPCConnector
 from goblet.infrastructures.cloudtask import CloudTaskQueue
+from goblet.infrastructures.pubsub import PubSubTopic
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.getLevelName(os.getenv("GOBLET_LOG_LEVEL", "INFO")))
@@ -40,6 +41,7 @@ SUPPORTED_INFRASTRUCTURES = {
     "redis": Redis,
     "vpcconnector": VPCConnector,
     "cloudtaskqueue": CloudTaskQueue,
+    "pubsub_topic": PubSubTopic,
 }
 
 
@@ -128,12 +130,20 @@ class Goblet_Decorators:
             },
         )
 
-    def topic(self, topic, **kwargs):
+    def pubsub_subscription(self, topic, **kwargs):
         """Pubsub topic trigger"""
         return self._create_registration_function(
             handler_type="pubsub",
             registration_kwargs={"topic": topic, "kwargs": kwargs},
         )
+
+    def topic(self, topic, **kwargs):
+        warn(
+            "This method is deprecated, use @app.pubsub_subscription",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return self.pubsub_subscription(topic, **kwargs)
 
     def cloudtasktarget(self, name, **kwargs):
         """CloudTask trigger"""
@@ -238,6 +248,14 @@ class Goblet_Decorators:
         kwargs["config"] = config
         return self._register_infrastructure(
             handler_type="cloudtaskqueue",
+            kwargs={"name": name, "config": config, "kwargs": kwargs},
+        )
+
+    def pubsub_topic(self, name, config=None, **kwargs):
+        """PubSub Topic Infrastructure"""
+        kwargs["config"] = config
+        return self._register_infrastructure(
+            handler_type="pubsub_topic",
             kwargs={"name": name, "config": config, "kwargs": kwargs},
         )
