+++ goblet/tests/data/http/pubsub-deploy/delete-v1-projects-goblet-topics-test_1.json
@@ -1,0 +1,4 @@
+{
+    "headers": {},
+    "body": {}
+}
