+++ goblet/tests/data/http/pubsub-deploy/put-v1-projects-goblet-topics-test_1.json
@@ -1,0 +1,6 @@
+{   
+    "headers": {},
+    "body": {
+        "name": "projects/goblet/topics/test"
+    }
+}
