+++ goblet/tests/data/http/pubsub-update/patch-v1-projects-goblet-topics-test_1.json
@@ -1,0 +1,7 @@
+{   
+    "headers": {},
+    "body": {
+        "name": "projects/goblet/topics/test",
+        "messageRetentionDuration": "3600s"
+    }
+}
